move summary stat
